Start cleaning out the Disk tests to their own module.
